Rearranging the deck chairs - sorting the output.
Sort the connections be time order (using the first transmission as a
start) and then sorting the transmissions within each connection too.

Due to the way the data is coming together we were seeing quite a
disconnect between the sides of the communications and this hopefuly
evens that out.

I do wonder if it might be better using a different library to
reconstruct the data as the reader idea is quite neat, but does mean you
lose a lot of the sequence information.