Add a --no-sort flag to not sort the data by times.
This allows you to see the order in which the data was actually
processed rather than the order it was sent over the wire.  This isn't
going to be helpful for interpretting the data, but is useful for
debugging anomalies in processing.

This may turn out to be something that only makes sense for a short
while as the data really needs to be processed in the real order the
data went over the wire, but the libraries for processing it don't emit
it in strictly that order.  If we fix that, then this flag probably
becomes redundant.