Fix an oddity with constructing the raw data field.
Because we try to write multiple times when we aren't sure if we have
all of the packet, we can end up having written too much.  This makes
the rawdata we've grabbed from the multiwriter have duplicate data and
general garbling.

Reordering at least means we don't record the data from the errors we
get from that process, but it does mean we don't get them when we legit
have an error.

I also wonder if there is another source of weirdness in that we might
only take a chunk of the data, and the multiwriter might not deal with
that?  We potentially need to do our own version of a multiwriter that
deals with the odd process we use fully.