Start to deal with multiple packets within a compressed block of data.
This is starting to feel like a dead end.  I think the fundamental
problem we have is that we have multiple compressed packets making up a
single uncompressed one.

It may be best if we go through the compressed stream and then
uncompress it before doing the normal processing.  This is less
efficient in terms of memory use, but is probably simpler for a start.