Rearchitect decompression
The compressed packets can contain chunks of a single packet, so we need
to push chunks of that onto the buffer so we can reconstruct that packet
from the data that was compressed.

I suspect that if we somehow were given multiple packets within a
compressed packet this might go wrong, but not sure if that's really a
thing.