Prevent crash if same effective connection comes around again.
If we encounter the same src-dest address combination we were crashing
because our builder had nil'd some state that was needed during
construction.  Now if we have a completed builder in the map and have
encountered that connection again we replace it so that we can deal with
the next connection.

Now we're pushing out the results via channels that would work fine.